HIVE-27166 : Introduce Apache Commons DBUtils to handle boilerplate code
