+++ standalone-metastore/metastore-server/pom.xml
@@ -79,6 +79,10 @@
       <artifactId>commons-dbcp2</artifactId>
     </dependency>
     <dependency>
+      <groupId>commons-dbutils</groupId>
+      <artifactId>commons-dbutils</artifactId>
+    </dependency>
+    <dependency>
       <groupId>io.dropwizard.metrics</groupId>
       <artifactId>metrics-core</artifactId>
     </dependency>
