+++ standalone-metastore/pom.xml
@@ -62,6 +62,7 @@
     <commons-lang3.version>3.9</commons-lang3.version>
     <commons-logging.version>1.1.3</commons-logging.version>
     <commons-dbcp2.version>2.7.0</commons-dbcp2.version>
+    <commons-dbutils.version>1.7</commons-dbutils.version>
     <datasketches.version>1.1.0-incubating</datasketches.version>
     <datanucleus-api-jdo.version>5.2.8</datanucleus-api-jdo.version>
     <datanucleus-core.version>5.2.10</datanucleus-core.version>
@@ -272,6 +273,11 @@
         <groupId>org.apache.commons</groupId>
         <artifactId>commons-dbcp2</artifactId>
         <version>${commons-dbcp2.version}</version>
+      </dependency>
+      <dependency>
+        <groupId>commons-dbutils</groupId>
+        <artifactId>commons-dbutils</artifactId>
+        <version>${commons-dbutils.version}</version>
       </dependency>
       <dependency>
         <groupId>org.apache.logging.log4j</groupId>
